db_api/requirements.txt
brotli==1.0.7
cryptography==2.9.2
Cython==0.29.20
fastapi==0.58.0
ipaddr==2.2.0
keyring==21.2.1
mock==4.0.2
mypy_extensions==0.4.3
numpy==1.18.5
ordereddict==1.1
protobuf==3.12.2
pyOpenSSL==19.1.0
python-dotenv==0.13.0
simplejson==3.17.0
SQLAlchemy==1.3.17
wincertstore==0.2
